Setup initial ApiClient and AuthService routes
